ServerSide/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.26
pymongo==4.6.0
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
